migrate tags to forumTags, remove util tag
