+++ backend/fithub/forum/serializers.py
@@ -1,15 +1,14 @@
 # forum/serializers.py
 from rest_framework import serializers
-from forum.models import ForumPost
-from utils.models import Tag
+from forum.models import ForumPost, ForumTag
 
-class TagSerializer(serializers.ModelSerializer):
+class ForumTagSerializer(serializers.ModelSerializer):
     class Meta:
-        model = Tag
+        model = ForumTag
         fields = ['name']
 
 class ForumPostSerializer(serializers.ModelSerializer):
-    tags = TagSerializer(many=True, read_only=True)
+    tags = ForumTagSerializer(many=True, read_only=True)
     tag_names = serializers.ListField(
         child=serializers.CharField(),
         write_only=True
@@ -29,7 +28,7 @@ class ForumPostSerializer(serializers.ModelSerializer):
 
     def validate_tag_names(self, value):
         """Validate that all tag names are valid choices."""
-        valid_tag_names = [choice.value for choice in Tag.TagChoices]
+        valid_tag_names = [choice.value for choice in ForumTag.TagChoices]
         invalid_tags = [name for name in value if name not in valid_tag_names]
 
         if invalid_tags:
@@ -38,7 +37,7 @@ class ForumPostSerializer(serializers.ModelSerializer):
                 f"Valid tags are: {', '.join(valid_tag_names)}."
             )
         return value
-    
+
     def create(self, validated_data):
         tag_names = validated_data.pop('tag_names', [])
         user = self.context['request'].user
@@ -46,6 +45,6 @@ class ForumPostSerializer(serializers.ModelSerializer):
         post = ForumPost.objects.create(author=user, **validated_data)
 
         # Assign tags based on tag names
-        tags = Tag.objects.filter(name__in=tag_names)
-        post.tags.set(tags)
+        forum_tags = ForumTag.objects.filter(name__in=tag_names)
+        post.tags.set(forum_tags)
         return post
