+++ backend/fithub/forum/urls.py
@@ -1,11 +1,11 @@
 # forum/urls.py
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from forum.views import ForumPostViewSet, TagViewSet
+from forum.views import ForumPostViewSet, ForumTagViewSet
 
 router = DefaultRouter()
 router.register(r'posts', ForumPostViewSet, basename='forumpost') # For forum post endpoints
-router.register(r'tags', TagViewSet, basename='tag') # For tags and points
+router.register(r'tags', ForumTagViewSet, basename='forumtag') # For tags and points
 
 urlpatterns = [
     path('', include(router.urls)),
